Modulize scan context manager, add roucount for olaptable etc.
